Modifications about package architectures
The git common organization of our packages is reviewed adding a "wip"
branch for development ("dev" must now be the mirror of master with
infos for developers).

The default "gitignore" is modified to follow the new global package
common architecture and the "gitignore-dev" is deleted as it was not
really usefull.

Correction of a typo, the plural of "media" in english is ... "media"
(renaming the whole directory).